requirements.txt
websockets==8.1
websocket-client==0.57.0
requests==2.23.0